moved PublicCalendarAPIMiddleware to request
